+++ src/blocks.js
@@ -11064,11 +11064,13 @@ InputSlotMorph.prototype.typesMenu = function () {
         list : ['list']
     };
     if (SpriteMorph.prototype.enableFirstClass) {
+        dict.agent = ['agent'];
         dict.sprite = ['sprite'];
         dict.stage = ['stage'];
     }
     dict.costume = ['costume'];
     dict.sound = ['sound'];
+    dict.script = ['script'];
     dict.command = ['command'];
     dict.reporter = ['reporter'];
     dict.predicate = ['predicate'];